LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/data-url-iframe-expected.txt

CONSOLE MESSAGE: Origin null is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/top.txt due to access control checks.


PASS fetching "top.txt" without ACAO should be rejected.
PASS fetching "top.txt" with CORS allowing null origin should be allowed.
PASS fetching data url script should be allowed.
